ioloop: Fix hanging outgoing TCP connections
TCP handles are not added to the IO loop until a second call to
uv__io_poll(). Collect the uv_handles with a file descriptor and
if they changed, ensure we call uv_run() again by doing it once
more, or on the next IO loop iteration.

Add a test that does 100 HTTP requests via two zeek instances.