types: Support setting Zeek record fields
This allows to implement hooks or event handlers in JS that
actually set fields of records.

    // File:: extend-conn.zeek
    @load base/protocols/conn

    @load ./extend-conn.js

    redef record Conn::Info += {
      js_count: count &log &optional;
    };

    // File extend-conn.js
    var count = 0;
    zeek.on('connection_state_remove', function(c) {
      c.conn.js_count = ++count;
    });

There are some caveats/quirks for complex types like vectors or sets,
but for primitive value types this should allow some nice demonstrations
for simple use cases without the need to resort back to Zeek scripts.