cmake: Improve for builtin plugin
Avoid finding libnode in the default system directories if an
explicit NODEJS_ROOT_DIR was set. That was always confusing,
extract the version and include libuv library in NODEJS_LIBRARIES.
If node was built with libuv as a shared component, ensure the
plugin has a link dependency on libuv as well so that it is
propagated to the Zeek executable when builtin.