Plugin: Tweak termination decision
Previously, in reading_traces mode, the IO loop would've been
kept alive a bit longer than without the plugin, causing time
to be forwarded by Zeek 5.2 and before by the event manager.

This has changed in Zeek recently and caused a bit of baseline
changes. Update and add a test to ensure Zeek and JavaScript
execute events in the same order.