Types: Cache wrapped objects
Do not create a new object instance if we did so previously and it
still exists in the Heap. This seems to be more performant with a
small amount of live Zeek objects wrapped by JavaScript objects
compared to re-instantiating new ones.

The map lookup may become significant if there are thousands of
objects held live though.