Types: Minor cleanup and docs for multi-index table handling
Calling console.log() on a JavaScript object backing a Zeek table with multiple
indices, node 19.8 triggers a crash due to the exception happening trying to
string format the table. I'm not quite sure what to do about it other than
advising not to do console.log() on such tables. This change adds a bit of
commentary to the code and cleanups, but otherwise does not change anything
in this regard.