Plugin: Avoid creating Exprs at runtime
We've previously created a CallExpr for every zeek.invoke() call to
provide caller information location to Zeek. This probably was never
a good idea. Proposed changes on the Zeek side are copying away the
instance pointer and triggered ref counting errors as well as memory
leaks if we continue to do it this way.

Relates to zeek/zeek#3433