Enable `cmake/FindNodejs.cmake` to find more libnode versions
We would previously hardcode a list of Node module versions which we
would "support" and find automatically; we however did and do not really
vet these versions. This meant that we were preptually unable to find
newer Node library versions until we patch the CMake module.

With this patch we automatically populate a list of current and future
Node module versions we recognize. Compilation might still fail down the
line, but this blanket discovery should simplify working with newer Node
versions nevertheless.