Instance: Offload JavaScript invocations to executor thread
As discussed in zeek/zeek#4239, sometimes the Zeek event queue is
drained with fiber stacks activated. This stack switching is not
something that V8 stack checking and garbage-collection is prepared
for. However, V8 has native support for threads. Offload JavaScript
execution to a dedicated thread to avoid executing V8 code from the
main thread on an alternative stack.

Fixes zeek/zeek#4239.