Node.js: Install uv_timer to imitate being alive
This prevents libuv considering its loop as inactive and not fully
processing it, even if Zeek keeps running. This issue was observed
when active signal handlers wouldn't be processed once there was no
alive handle registered with the loop and Zeek would start processing
the ZeekJS IO source over and over as a signal was pending.

Closes #11