Nodejs: Force GC during Done() and improve cleanup
When running with ASAN, StringVals continued to be referenced at shutdown
because V8 wouldn't run GC and some string resources would continue to
exists holding references. This is unfortunate for testing.

Rework Done() to force garbage collection using --expose-gc. Also call
some global shutdown functions for V8 Node.js and libuv.

Modify some of the tests to disable leak checking, or call terminate()
instead of exit. The issue here is that calling process.exit() doesn't
go through Plugin::Done() and so the cleanup does not run. This could
probably be improved with a custom exit handler and splitting Done()
into a helper for cleaning up, but that's for another time.

With these changes, it's now possible to remove the suppression
configuration altogether, so I think that was worth it.