Nodejs: Fix args passed as copy for hook handler
Using a reference / pointer avoids refing all the ValPtr instances
captured into the lambda. We do not want to do that as it results in
races during destruction of the packaged task as the main thread continues
as soon as the promise was resolved, but that's potentially before the
packaged task (including any ValPtr instances) within the executor thread
was destructed properly.

This should also be more efficient.

Might make sense to re-think that hand-off, but this seems to fix
the issue from zeek/zeek#5106.